+++ src/main/java/com/wedevol/springdatacouchbase/core/dao/doc/ProductDoc.java
@@ -1,0 +1,109 @@
+package com.wedevol.springdatacouchbase.core.dao.doc;
+
+import java.io.Serializable;
+import org.springframework.data.couchbase.core.mapping.Document;
+import org.springframework.data.couchbase.core.mapping.id.GeneratedValue;
+import org.springframework.data.couchbase.core.mapping.id.GenerationStrategy;
+import com.couchbase.client.java.repository.annotation.Field;
+import com.couchbase.client.java.repository.annotation.Id;
+
+/**
+ * Represents a Product doc from Couchbase. The doc key is generated using a unique number.
+ *
+ * @author Charz++
+ */
+
+@Document
+public class ProductDoc implements Serializable {
+
+  private static final long serialVersionUID = 3072475211055736282L;
+
+  @Id
+  @GeneratedValue(strategy = GenerationStrategy.UNIQUE)
+  private String id;
+
+  @Field
+  private Long name;
+  @Field
+  private String description;
+  @Field
+  private Integer quantity;
+
+  public ProductDoc() {}
+
+  public String getId() {
+    return id;
+  }
+
+  public void setId(String id) {
+    this.id = id;
+  }
+
+  public Long getName() {
+    return name;
+  }
+
+  public void setName(Long name) {
+    this.name = name;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
+  }
+
+  public Integer getQuantity() {
+    return quantity;
+  }
+
+  public void setQuantity(Integer quantity) {
+    this.quantity = quantity;
+  }
+
+  @Override
+  public int hashCode() {
+    final int prime = 31;
+    int result = 1;
+    result = prime * result + ((description == null) ? 0 : description.hashCode());
+    result = prime * result + ((id == null) ? 0 : id.hashCode());
+    result = prime * result + ((name == null) ? 0 : name.hashCode());
+    result = prime * result + ((quantity == null) ? 0 : quantity.hashCode());
+    return result;
+  }
+
+  @Override
+  public boolean equals(Object obj) {
+    if (this == obj)
+      return true;
+    if (obj == null)
+      return false;
+    if (getClass() != obj.getClass())
+      return false;
+    ProductDoc other = (ProductDoc) obj;
+    if (description == null) {
+      if (other.description != null)
+        return false;
+    } else if (!description.equals(other.description))
+      return false;
+    if (id == null) {
+      if (other.id != null)
+        return false;
+    } else if (!id.equals(other.id))
+      return false;
+    if (name == null) {
+      if (other.name != null)
+        return false;
+    } else if (!name.equals(other.name))
+      return false;
+    if (quantity == null) {
+      if (other.quantity != null)
+        return false;
+    } else if (!quantity.equals(other.quantity))
+      return false;
+    return true;
+  }
+
+}