+++ src/main/java/com/wedevol/springdatacouchbase/core/exception/ErrorType.java
@@ -6,7 +6,8 @@ package com.wedevol.springdatacouchbase.core.exception;
  * @author charz
  */
 public enum ErrorType {
-  USER_NOT_FOUND(404, "User not found"), USER_ALREADY_EXISTS(400, "User already exists");
+  USER_NOT_FOUND(1, "User not found"), USER_ALREADY_EXISTS(2, "User already exists"), CAR_NOT_FOUND(3,
+      "Car not found"), PRODUCT_NOT_FOUND(4, "Product not found");
 
   private final int code;
   private final String message;