+++ src/main/java/com/wedevol/springdatacouchbase/core/service/CarService.java
@@ -1,0 +1,16 @@
+package com.wedevol.springdatacouchbase.core.service;
+
+import com.wedevol.springdatacouchbase.core.dao.doc.CarDoc;
+
+/**
+ * Interface for the Car Service Implementation
+ *
+ * @author Charz++
+ */
+
+public interface CarService {
+
+  CarDoc findByKey(Long number, String manufacturer);
+
+  void create(CarDoc car);
+}