+++ src/main/java/com/wedevol/springdatacouchbase/core/service/ProductService.java
@@ -1,0 +1,16 @@
+package com.wedevol.springdatacouchbase.core.service;
+
+import com.wedevol.springdatacouchbase.core.dao.doc.ProductDoc;
+
+/**
+ * Interface for the Product Service Implementation
+ *
+ * @author Charz++
+ */
+
+public interface ProductService {
+
+  ProductDoc findById(String id);
+
+  void create(ProductDoc product);
+}